+++ src/game/collision.cpp
@@ -1155,7 +1155,7 @@ int CCollision::IntersectNoLaser(vec2 Pos0, vec2 Pos1, vec2 *pOutCollision, vec2
 
 	for(int i = 0, id = std::ceil(d); i < id; i++)
 	{
-		float a = (int)i / d;
+		float a = i / d;
 		vec2 Pos = mix(Pos0, Pos1, a);
 		int Nx = clamp(round_to_int(Pos.x) / 32, 0, m_Width - 1);
 		int Ny = clamp(round_to_int(Pos.y) / 32, 0, m_Height - 1);
